+++ p2p/common.h
@@ -43,8 +43,8 @@ struct data_header {
 };
 
 
-void initialize_srvr(int* sockfd, const int* domain, const uint16_t* port, const uint32_t* address);
-void initialize_clnt(int* sockfd, const int* domain, const uint16_t* port, const uint32_t* address, struct sockaddr_in* sockaddr);
+void initialize_srvr(int* sockfd, const short* domain, const uint16_t* port, const uint32_t* address);
+void initialize_clnt(int* sockfd, const short* domain, const uint16_t* port, const uint32_t* address, struct sockaddr_in* sockaddr);
 void recv_message(const int* sockfd, void* data, const size_t data_len, int flags, struct sockaddr_in* sockaddr);
 void send_message(const int* sockfd, const void* data, const size_t data_len, int flags, const struct sockaddr_in* sockaddr);
